Clean up filter functions
- consolidate AbstractFilterF and AbstractFFTFilterF
- make naming consistent (use 'paddedInput' instead of
'raiExtendedInput'
