+++ src/main/java/net/imagej/ops/filter/AbstractFFTFilterF.java
@@ -33,16 +33,21 @@ import net.imagej.ops.filter.fft.CreateOutputFFTMethods;
 import net.imagej.ops.filter.pad.PadInputFFTMethods;
 import net.imagej.ops.filter.pad.PadShiftKernelFFTMethods;
 import net.imagej.ops.special.computer.BinaryComputerOp;
+import net.imagej.ops.special.function.AbstractBinaryFunctionOp;
 import net.imagej.ops.special.function.BinaryFunctionOp;
 import net.imagej.ops.special.function.Functions;
 import net.imagej.ops.special.function.UnaryFunctionOp;
 import net.imglib2.Dimensions;
 import net.imglib2.FinalDimensions;
 import net.imglib2.RandomAccessibleInterval;
+import net.imglib2.outofbounds.OutOfBoundsFactory;
 import net.imglib2.type.NativeType;
+import net.imglib2.type.Type;
 import net.imglib2.type.numeric.ComplexType;
 import net.imglib2.type.numeric.RealType;
 import net.imglib2.type.numeric.complex.ComplexFloatType;
+import net.imglib2.type.numeric.real.FloatType;
+import net.imglib2.util.Util;
 
 import org.scijava.plugin.Parameter;
 
@@ -57,14 +62,50 @@ import org.scijava.plugin.Parameter;
  * @param <C>
  */
 public abstract class AbstractFFTFilterF<I extends RealType<I>, O extends RealType<O> & NativeType<O>, K extends RealType<K>, C extends ComplexType<C> & NativeType<C>>
-	extends AbstractFilterF<I, O, K, C>
+	extends
+	AbstractBinaryFunctionOp<RandomAccessibleInterval<I>, RandomAccessibleInterval<K>, RandomAccessibleInterval<O>>
 {
+
+	/**
+	 * Border size in each dimension. If null default border size will be
+	 * calculated and added.
+	 */
+	@Parameter(required = false)
+	private long[] borderSize = null;
+
+	/**
+	 * Defines the out of bounds strategy for the extended area of the input
+	 */
+	@Parameter(required = false)
+	private OutOfBoundsFactory<I, RandomAccessibleInterval<I>> obfInput;
+
+	/**
+	 * Defines the out of bounds strategy for the extended area of the kernel
+	 */
+	@Parameter(required = false)
+	private OutOfBoundsFactory<K, RandomAccessibleInterval<K>> obfKernel;
+
+	/**
+	 * The output type. If null a default output type will be used.
+	 */
+	@Parameter(required = false)
+	private Type<O> outType;
 
 	/**
 	 * FFT type
 	 */
 	@Parameter(required = false)
 	private ComplexType<C> fftType;
+
+	/**
+	 * Op used to pad the input
+	 */
+	private BinaryFunctionOp<RandomAccessibleInterval<I>, Dimensions, RandomAccessibleInterval<I>> padOp;
+
+	/**
+	 * Op used to pad the kernel
+	 */
+	private BinaryFunctionOp<RandomAccessibleInterval<K>, Dimensions, RandomAccessibleInterval<K>> padKernelOp;
 
 	/**
 	 * Op used to create the complex FFTs
@@ -84,16 +125,16 @@ public abstract class AbstractFFTFilterF<I extends RealType<I>, O extends RealTy
 		/**
 		 * Op used to pad the input
 		 */
-		setPadOp((BinaryFunctionOp) Functions.binary(ops(),
-			PadInputFFTMethods.class, RandomAccessibleInterval.class,
-			RandomAccessibleInterval.class, Dimensions.class, true, getOBFInput()));
+		padOp = (BinaryFunctionOp) Functions.binary(ops(), PadInputFFTMethods.class,
+			RandomAccessibleInterval.class, RandomAccessibleInterval.class,
+			Dimensions.class, true, obfInput);
 
 		/**
 		 * Op used to pad the kernel
 		 */
-		setPadKernelOp((BinaryFunctionOp) Functions.binary(ops(),
+		padKernelOp = (BinaryFunctionOp) Functions.binary(ops(),
 			PadShiftKernelFFTMethods.class, RandomAccessibleInterval.class,
-			RandomAccessibleInterval.class, Dimensions.class, true));
+			RandomAccessibleInterval.class, Dimensions.class, true);
 
 		if (fftType == null) {
 			fftType = (ComplexType<C>) ops().create().nativeType(
@@ -109,13 +150,73 @@ public abstract class AbstractFFTFilterF<I extends RealType<I>, O extends RealTy
 	}
 
 	/**
+	 * Create the output using the outFactory and outType if they exist. If these
+	 * are null use a default factory and type
+	 */
+	@SuppressWarnings("unchecked")
+	public RandomAccessibleInterval<O> createOutput(
+		RandomAccessibleInterval<I> input, RandomAccessibleInterval<K> kernel)
+	{
+
+		if (outType == null) {
+
+			// if the input type and kernel type are the same use this type
+			if (Util.getTypeFromInterval(input).getClass() == Util
+				.getTypeFromInterval(kernel).getClass())
+			{
+				Object temp = Util.getTypeFromInterval(input).createVariable();
+				outType = (Type<O>) temp;
+
+			}
+			// otherwise default to float
+			else {
+				Object temp = new FloatType();
+				outType = (Type<O>) temp;
+			}
+		}
+
+		return ops().create().img(input, outType.createVariable());
+	}
+
+	/**
 	 * create FFT memory, create FFT filter and run it
 	 */
 	@Override
-	public void computeFilter(final RandomAccessibleInterval<I> input,
-		final RandomAccessibleInterval<K> kernel,
-		RandomAccessibleInterval<O> output, long[] paddedSize)
+	public RandomAccessibleInterval<O> calculate(
+		final RandomAccessibleInterval<I> input,
+		final RandomAccessibleInterval<K> kernel)
 	{
+
+		RandomAccessibleInterval<O> output = createOutput(input, kernel);
+
+		final int numDimensions = input.numDimensions();
+
+		// 1. Calculate desired extended size of the image
+
+		final long[] paddedSize = new long[numDimensions];
+
+		if (borderSize == null) {
+			// if no border size was passed in, then extend based on kernel size
+			for (int d = 0; d < numDimensions; ++d) {
+				paddedSize[d] = (int) input.dimension(d) + (int) kernel.dimension(d) -
+					1;
+			}
+
+		}
+		else {
+			// if borderSize was passed in
+			for (int d = 0; d < numDimensions; ++d) {
+
+				paddedSize[d] = Math.max(kernel.dimension(d) + 2 * borderSize[d], input
+					.dimension(d) + 2 * borderSize[d]);
+			}
+		}
+
+		RandomAccessibleInterval<I> paddedInput = padOp.calculate(input,
+			new FinalDimensions(paddedSize));
+
+		RandomAccessibleInterval<K> paddedKernel = padKernelOp.calculate(kernel,
+			new FinalDimensions(paddedSize));
 
 		RandomAccessibleInterval<C> fftInput = createOp.calculate(
 			new FinalDimensions(paddedSize));
@@ -127,26 +228,52 @@ public abstract class AbstractFFTFilterF<I extends RealType<I>, O extends RealTy
 		// 'initialize' as we don't know the size yet, thus we can't create
 		// memory
 		// for the FFTs
-		filter = createFilterComputer(input, kernel, fftInput, fftKernel, output);
+		filter = createFilterComputer(paddedInput, paddedKernel, fftInput,
+			fftKernel, output);
 
-		filter.compute(input, kernel, output);
+		filter.compute(paddedInput, paddedKernel, output);
+
+		return output;
 	}
 
 	/**
 	 * This function is called after the RAIs and FFTs are set up and create the
 	 * frequency filter computer.
 	 * 
-	 * @param raiExtendedInput
-	 * @param raiExtendedKernel
+	 * @param paddedInput
+	 * @param paddedKernel
 	 * @param fftImg
 	 * @param fftKernel
 	 * @param output
 	 */
 	abstract public
 		BinaryComputerOp<RandomAccessibleInterval<I>, RandomAccessibleInterval<K>, RandomAccessibleInterval<O>>
-		createFilterComputer(RandomAccessibleInterval<I> raiExtendedInput,
-			RandomAccessibleInterval<K> raiExtendedKernel,
+		createFilterComputer(RandomAccessibleInterval<I> paddedInput,
+			RandomAccessibleInterval<K> paddedKernel,
 			RandomAccessibleInterval<C> fftImg, RandomAccessibleInterval<C> fftKernel,
 			RandomAccessibleInterval<O> output);
 
+	protected long[] getBorderSize() {
+		return borderSize;
+	}
+
+	protected OutOfBoundsFactory<I, RandomAccessibleInterval<I>> getOBFInput() {
+		return obfInput;
+	}
+
+	protected void setOBFInput(OutOfBoundsFactory<I, RandomAccessibleInterval<I>> objInput) {
+		this.obfInput = objInput;
+	}
+
+	protected OutOfBoundsFactory<K, RandomAccessibleInterval<K>> getOBFKernel() {
+		return obfKernel;
+	}
+
+	protected void setOBFKernel(OutOfBoundsFactory<K, RandomAccessibleInterval<K>> obfKernel) {
+		this.obfKernel = obfKernel;
+	}
+
+	protected Type<O> getOutType() {
+		return outType;
+	}
 }
