+++ src/main/java/net/imagej/ops/filter/convolve/ConvolveFFTF.java
@@ -77,13 +77,13 @@ public class ConvolveFFTF<I extends RealType<I> & NativeType<I>, O extends RealT
 	@Override
 	public
 		BinaryComputerOp<RandomAccessibleInterval<I>, RandomAccessibleInterval<K>, RandomAccessibleInterval<O>>
-		createFilterComputer(RandomAccessibleInterval<I> raiExtendedInput,
-			RandomAccessibleInterval<K> raiExtendedKernel,
+		createFilterComputer(RandomAccessibleInterval<I> paddedInput,
+			RandomAccessibleInterval<K> paddedKernel,
 			RandomAccessibleInterval<C> fftImg, RandomAccessibleInterval<C> fftKernel,
 			RandomAccessibleInterval<O> output)
 	{
-		return Computers.binary(ops(), ConvolveFFTC.class, output, raiExtendedInput,
-			raiExtendedKernel, fftImg, fftKernel);
+		return Computers.binary(ops(), ConvolveFFTC.class, output, paddedInput,
+			paddedKernel, fftImg, fftKernel);
 	}
 
 }
